Syntax: Support path expressions in uRust syntax
Change the uRust syntax grammar to directly support path expressions of the form
Foo::Bar. After parsing, these are emitted as `_urust_path_string_identifier`
entries in the AST.

In the shallow embedding of uRust, such path identifiers are now looked up in a
dictionary. Change `notation_nano_rust` and friends to add new notation to that
dictionary, if we are registering a path identifier.

This reduces the number of grammar entries produced when using
`notation_nano_rust`. Large grammars can cause slowdowns when importing theories.

Signed-off-by: Ike Mulder <ikemul@amazon.com>