Misc: Add plugin that removes the quadratic nr of elim rules for new datatypes
When defining a new datatype with n constructors, a quadratic number of
elimination rules are created and added to the ruleset used by `clarify`.
Unfortunately, Isabelle does not seem to be very performant when merging
large amounts of elimination rules during theory imports. See also [1].

As a stopgap, add a plugin in Misc which removes the elimination rules for
every newly defined datatype.

Take the time to add `Result` to the `Misc/ROOT` file, which was missing.

[1]: https://isabelle.zulipchat.com/#narrow/channel/247541-Mirror.3A-Isabelle-Users-Mailing-List/topic/.5Bisabelle.5D.20Merging.20elimination.20rules.20of.20Claset.20is.20slow.20-.20.2E.2E.2E/near/525149986

Signed-off-by: Ike Mulder <ikemul@amazon.com>