I/Q: Preserve MCP request ID
Previously, I/Q would reply with ID "1.0" (as a string) to a request
with ID 1 (as an integer), tripping some MCP clients.

This commit fixes this by having I/Q simply preserve the ID as-is,
regardless of its type (which may be a string, an whole integer, or a
fractional integer).

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>