I/Q: Move response+error formatting logic into separate functions
Previously, each tool handler would return the final JSON reply, including
the JSON-RPC header.

To avoid code duplication and simplify changes in the response formatting,
this commit hoists out the shared formatting logic into helper functions
`formatErrorResponse` and `formatSuccessResponse`. The core handlers now
return `Map` objects. Additionally, a helper wrapToolCallResult is provided
which can be used to uniformly change the formatting of tool call responses.
For now, this is the identity, but it will be adjusted in a subsequent
commit.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>