Fix match on numerals not supporting zero and one
It turns out that the literals '0' and '1' are not parsed like other numerals.
Instead, zero and one have their own notation, and this means the approach
from #51 does not yet work for matching on zero and one.

Therefore, add explicit clauses to the syntax for matching on zero and one.
Make sure to add tests for such matches.

Fixes: #51

Signed-off-by: Ike Mulder <ikemul@amazon.com>